requirements.txt
line-bot-sdk==3.11.0
Flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.7.1
langchain-google-genai
requests~=2.31.0
beautifulsoup4==4.12.3
pydantic~=2.8.2
pymongo[srv]
google==3.0.0
uvicorn==0.30.1
